Removing rank code
Heroku only offers daily and hourly cron jobs and
I don't think either would be suitable for
updating rank. We could pay for a worker but I am
not sure if my company wants to pay for that right
now. I would like to revisit this later and use
something like redis.